Loglikelihood adjustments for fitted models

This function adjusts the loglikelihood of fitted model objects based on Chandler and Bate (2007). It is a generic function for different types of models, which are listed in Supported models. This section also contains links to function-specific help pages.

adj_loglik(x, cluster = NULL, use_vcov = TRUE, use_mle = TRUE, ...)

Arguments

x
A supported fitted model object, see Supported models

cluster
A vector or factor indicating the cluster the corresponding loglikelihood contribution belongs to. It is required to have the same length as the vector returned by logLik_vec(). If cluster is not supplied or NULL, then it is assumed that each observation forms its own cluster.

use_vcov
A logical scalar. By default, the vcov() method for x is used to estimate the Hessian of the independence loglikelihood, if the function exists. Otherwise, or if use_vcov = FALSE, H is estimated using stats::optimHess() inside chandwich::adjust_loglik().

use_mle
A logical scalar. By default, the MLE from x is taken as given, and is not reestimated. By setting use_mle to FALSE, the parameters are reestimated in the function chandwich::adjust_loglik() using stats::optim().This feature is currently for development purposes only, may return misleading/false results and may be removed without notice.

...
Further arguments to be passed to sandwich::meatCL() if cluster is defined, if cluster = NULL, they are passed into sandwich::meat().

Value

An object of class "chantrics" inheriting from class "chandwich". See the documentation provided with chandwich::adjust_loglik().

Details

If use_vcov = TRUE, the current default, the function will test whether a vcov S3 method exists for x, and will take the variance-covariance matrix from there. Otherwise, or if use_vcov = FALSE the variance-covariance matrix of the MLE is estimated inside chandwich::adjust_loglik() using stats::optimHess().

Supported models

glm

glm.nb

Available methods

"chantrics" objects have the following methods available to them:

AIC

anova

alrtest - Adjusted Likelihood ratio tests

coef

lmtest::coeftest - \(z\) tests for all coefficients

confint and plot.confint - confidence intervals for all coefficients, and diagnostics plots for confint().

conf_intervals - enhanced confidence interval reports

conf_region - two-dimensional confidence regions

df.residual

fitted

logLik

logLik_vec

nobs

plot

predict

print

residuals

summary

terms

vcov

Examples

See the model-specific pages in the supported models section.

References

R. E. Chandler and S. Bate, Inference for clustered data using the independence loglikelihood, Biometrika, 94 (2007), pp. 167–183. doi: 10.1093/biomet/asm015 .

See also