Loglikelihood adjustments for pscl::hurdle fits

Adjust the loglikelihood and the standard errors of a fitted pscl::hurdle() model.

Details

Describe zeroinfl models here

Note that the pscl::hurdle() model has to be run with the option x = TRUE in order for the adjustment to execute properly. Also note that the functions residuals.chantrics() and fitted.chantrics() are currently disabled for hurdle models. Additionally, sequential anova.chantrics() are not available

Supported families

Within each family, any link function should work.

geometric

poisson

negbin

binomial (for the zero mass distribution only)

Examples

# hurdle model from AER, pg. 139-140
library(pscl)

#> Classes and Methods for R developed in the
#> Political Science Computational Laboratory
#> Department of Political Science
#> Stanford University
#> Simon Jackman
#> hurdle and zeroinfl functions by Achim Zeileis

data("RecreationDemand", package = "AER")
rd_hurdle <- hurdle(trips ~ . | quality + income,
  data = RecreationDemand,
  dist = "negbin", x = TRUE
)
summary(rd_hurdle)

#> 
#> Call:
#> hurdle(formula = trips ~ . | quality + income, data = RecreationDemand, 
#>     dist = "negbin", x = TRUE)
#> 
#> Pearson residuals:
#>     Min      1Q  Median      3Q     Max 
#> -1.6097 -0.2071 -0.1845 -0.1642 12.1114 
#> 
#> Count model coefficients (truncated negbin with log link):
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  0.84193    0.38278   2.200  0.02784 *  
#> quality      0.17170    0.07234   2.374  0.01762 *  
#> skiyes       0.62236    0.19013   3.273  0.00106 ** 
#> income      -0.05709    0.06452  -0.885  0.37629    
#> userfeeyes   0.57634    0.38508   1.497  0.13448    
#> costC        0.05707    0.02169   2.632  0.00850 ** 
#> costS       -0.07752    0.01155  -6.713  1.9e-11 ***
#> costH        0.01237    0.01490   0.830  0.40640    
#> Log(theta)  -0.53031    0.26114  -2.031  0.04228 *  
#> Zero hurdle model coefficients (binomial with logit link):
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -2.76631    0.36230  -7.635 2.25e-14 ***
#> quality      1.50291    0.10032  14.981  < 2e-16 ***
#> income      -0.04467    0.07853  -0.569    0.569    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 
#> Theta: count = 0.5884
#> Number of iterations in BFGS optimization: 18 
#> Log-likelihood: -765.1 on 12 Df

# adjust model
adj_loglik(rd_hurdle)

#> 
#>  Call: chandwich::adjust_loglik(loglik = logLik_f, fitted_object = x,      p = p, init = init, par_names = par_names, name = paste(name_pieces,          collapse = "_"), mle = mle, H = H, V = V) 
#>  
#> Attributes, access using attr(., "name") or attributes(.)$name 
#>  [1] "adjSE"          "adjVC"          "C_cholesky"     "C_spectral"    
#>  [5] "call"           "chantrics_args" "class"          "cluster"       
#>  [9] "formula"        "free_pars"      "full_par_names" "HA"            
#> [13] "HI"             "loglik"         "loglik_args"    "loglikVecMLE"  
#> [17] "max_loglik"     "MLE"            "name"           "nobs"          
#> [21] "p_current"      "p_full"         "par_names"      "res_MLE"       
#> [25] "SE"             "theta"          "VC"